altDel {MPRADesignGenerator}	R Documentation

CREATE ALTERNATIVE SEQUENCES FOR DELETIONS

Description

Delete specified number of bases from ref seq. (Counts the number of bases in deletion string, removes bases starting at 73)

Usage

altDel(reference_sequence, deletion_bases)

Arguments

145bp
reference sequence, string containing bases to delete.

Value

alt sequence with specified bases removed (counts the number of bases in string, removes bases starting at 73)

[Package MPRADesignGenerator version 0.0.0.9000 Index]