checkDigest {MPRADesignGenerator}	R Documentation

CHECK FOR ALL THREE DIGESTION SITES

Description

Check all sequences in a dataframe for digestion sites (SfiI, KpnI, XbaI) (column with sequences tested should be named short_seq)

Usage

checkDigest(df_to_test)

Arguments

dataframe
with sequences to test

Value

new dataframe containing sequences which had a digestion site

[Package MPRADesignGenerator version 0.0.0.9000 Index]