fixSfiI {MPRADesignGenerator}	R Documentation

FIX SfiI DIGESTION SITE

Description

Changes SfiI site (GGCCNNNNNGGCC -> GCGCNNNNNGGCC) Will only first first digestion site in sequence.

Usage

fixSfiI(sequence, coord1, coord2)

Arguments

145bp
sequence containing site, coordinate of first base of restriction site, coordinate of second base of restriction site.

Value

corrected sequence

[Package MPRADesignGenerator version 0.0.0.9000 Index]